Create a helper file for font-related functions
