+++ lib/clean-assets/android.js
@@ -1,55 +1,20 @@
 const fs = require('fs-extra');
 const path = require('path');
 const OpenType = require('opentype.js');
-const slugify = require('slugify');
 const { fontTypes } = require('../helper');
-const { XMLParser, XMLBuilder } = require('fast-xml-parser');
-const { sync: globSync } = require('glob');
-
-const REACT_FONT_MANAGER_JAVA_IMPORT = 'com.facebook.react.views.text.ReactFontManager';
-
-function toArrayBuffer(buffer) {
-  const arrayBuffer = new ArrayBuffer(buffer.length);
-  const view = new Uint8Array(arrayBuffer);
-
-  for (let i = 0; i < buffer.length; i += 1) {
-    view[i] = buffer[i];
-  }
-
-  return arrayBuffer;
-}
-
-function normalizeString(str) {
-  return slugify(str, { lower: true }).replaceAll('-', '_');
-}
-
-function getProjectFilePath(rootPath, name) {
-  const filePath = globSync(path.join(rootPath, `android/app/src/main/java/**/${name}.java`))[0];
-  return filePath;
-}
-
-function getFontFamily(fontFamily, preferredFontFamily) {
-  const availableFontFamily = preferredFontFamily || fontFamily;
-  return availableFontFamily.en || Object.values(availableFontFamily)[0];
-}
-
-function getFontResFolderPath(rootPath) {
-  return path.join(rootPath, 'android/app/src/main/res/font');
-}
-
-function getXMLFontId(fontFileName) {
-  return `@font/${path.basename(fontFileName, path.extname(fontFileName))}`;
-}
-
-function getAddCustomFontMethodCall(fontName, fontId) {
-  return `ReactFontManager.getInstance().addCustomFont(this, "${fontName}", R.font.${fontId});`;
-}
-
-function removeLineFromJavaFile(javaFileData, stringToRemove) {
-  const lines = javaFileData.split('\n');
-  const updatedLines = lines.filter(line => !line.includes(stringToRemove));
-  return updatedLines.join('\n');
-}
+const {
+  REACT_FONT_MANAGER_JAVA_IMPORT,
+  toArrayBuffer,
+  normalizeString,
+  getProjectFilePath,
+  getFontFamily,
+  getFontResFolderPath,
+  getXMLFontId,
+  getAddCustomFontMethodCall,
+  removeLineFromJavaFile,
+  xmlParser,
+  xmlBuilder,
+} = require('../font-assets-helper');
 
 module.exports = function cleanAssetsAndroid(assetFiles = [], config, options, fileExt) {
   // If the assets are not fonts, just remove them.
@@ -60,8 +25,6 @@ module.exports = function cleanAssetsAndroid(assetFiles = [], config, options, f
     return;
   }
 
-  const mainApplicationFilePath = getProjectFilePath(config.rootPath, 'MainApplication');
-
   const fontFamilyMap = {};
 
   assetFiles.forEach((file) => {
@@ -70,6 +33,9 @@ module.exports = function cleanAssetsAndroid(assetFiles = [], config, options, f
     const buffer = fs.readFileSync(fontFilePath);
     const font = OpenType.parse(toArrayBuffer(buffer));
 
+    // Build the font family's map, where each key is the font family name,
+    // and each value is a object containing all the font files related to that
+    // font family.
     const fontFamily = getFontFamily(font.names.fontFamily, font.names.preferredFamily);
     if (!fontFamilyMap[fontFamily]) {
       fontFamilyMap[fontFamily] = {
@@ -84,18 +50,8 @@ module.exports = function cleanAssetsAndroid(assetFiles = [], config, options, f
     });
   });
 
-  const xmlParser = new XMLParser({
-    ignoreAttributes: false,
-    isArray: tagName => tagName === 'font',
-  });
-
-  const xmlBuilder = new XMLBuilder({
-    format: true,
-    ignoreAttributes: false,
-    suppressEmptyNode: true,
-  });
-
   // Read MainApplication.java file.
+  const mainApplicationFilePath = getProjectFilePath(config.rootPath, 'MainApplication');
   let mainApplicationFileData = fs.readFileSync(mainApplicationFilePath).toString();
 
   Object.entries(fontFamilyMap).forEach(([fontFamilyName, fontFamilyData]) => {
