+++ lib/copy-assets/android.js
@@ -1,172 +1,24 @@
 const fs = require('fs-extra');
 const path = require('path');
 const OpenType = require('opentype.js');
-const slugify = require('slugify');
 const { fontTypes } = require('../helper');
-const { XMLParser, XMLBuilder } = require('fast-xml-parser');
-const { sync: globSync } = require('glob');
-
-const REACT_FONT_MANAGER_JAVA_IMPORT = 'com.facebook.react.views.text.ReactFontManager';
-
-function toArrayBuffer(buffer) {
-  const arrayBuffer = new ArrayBuffer(buffer.length);
-  const view = new Uint8Array(arrayBuffer);
-
-  for (let i = 0; i < buffer.length; i += 1) {
-    view[i] = buffer[i];
-  }
-
-  return arrayBuffer;
-}
-
-function normalizeString(str) {
-  return slugify(str, { lower: true }).replaceAll('-', '_');
-}
-
-function getProjectFilePath(rootPath, name) {
-  const filePath = globSync(path.join(rootPath, `android/app/src/main/java/**/${name}.java`))[0];
-  return filePath;
-}
-
-function getFontFamily(fontFamily, preferredFontFamily) {
-  const availableFontFamily = preferredFontFamily || fontFamily;
-  return availableFontFamily.en || Object.values(availableFontFamily)[0];
-}
-
-// eslint-disable-next-line consistent-return
-function getFontFallbackWeight(weight) {
-  if (weight <= 500) {
-    return Math.max(Math.floor(weight / 100) * 100, 100);
-  } else if (weight > 500) {
-    return Math.min(Math.ceil(weight / 100) * 100, 900);
-  }
-}
-
-function getFontResFolderPath(rootPath) {
-  return path.join(rootPath, 'android/app/src/main/res/font');
-}
-
-function getXMLFontId(fontFileName) {
-  return `@font/${path.basename(fontFileName, path.extname(fontFileName))}`;
-}
-
-function buildXMLFontObjectEntry(fontFile) {
-  return {
-    '@_app:fontStyle': fontFile.isItalic ? 'italic' : 'normal',
-    '@_app:fontWeight': fontFile.weight,
-    '@_app:font': getXMLFontId(fontFile.name),
-  };
-}
-
-function buildXMLFontObject(fontFiles) {
-  return {
-    '?xml': {
-      '@_version': '1.0',
-      '@_encoding': 'utf-8',
-    },
-    'font-family': {
-      '@_xmlns:app': 'http://schemas.android.com/apk/res-auto',
-      font: fontFiles.map(buildXMLFontObjectEntry),
-    },
-  };
-}
-
-function getAddCustomFontMethodCall(fontName, fontId) {
-  return `ReactFontManager.getInstance().addCustomFont(this, "${fontName}", R.font.${fontId});`;
-}
-
-function addImportToJavaFile(javaFileData, importToAdd) {
-  const importRegex = new RegExp(`import\\s+${importToAdd};`, 'gm');
-  const existingImport = importRegex.exec(javaFileData);
-
-  if (existingImport) {
-    return javaFileData;
-  }
-
-  const packageRegex = /package\s+[\w.]+;/;
-  const packageMatch = packageRegex.exec(javaFileData);
-
-  let insertPosition = 0;
-
-  if (packageMatch) {
-    insertPosition = packageMatch.index + packageMatch[0].length;
-  }
-
-  return `${javaFileData.slice(0, insertPosition)}\n\nimport ${importToAdd};${javaFileData.slice(insertPosition)}`;
-}
-
-function insertLineInJavaClassMethod(
-  javaFileData,
-  targetClass,
-  targetMethod,
-  codeToInsert,
-  lineToInsertAfter,
-) {
-  const classRegex = new RegExp(`class\\s+${targetClass}(\\s+extends\\s+\\S+)?(\\s+implements\\s+\\S+)?\\s*\\{`, 'gm');
-  const classMatch = classRegex.exec(javaFileData);
-
-  if (!classMatch) {
-    console.error(`Class ${targetClass} not found.`);
-    return javaFileData;
-  }
-
-  const methodRegex = new RegExp(`(public|protected|private)\\s+(static\\s+)?\\S+\\s+${targetMethod}\\s*\\(`, 'gm');
-  let methodMatch = methodRegex.exec(javaFileData);
-
-  while (methodMatch) {
-    if (methodMatch.index > classMatch.index) {
-      break;
-    }
-    methodMatch = methodRegex.exec(javaFileData);
-  }
-
-  if (!methodMatch) {
-    console.error(`Method ${targetMethod} not found in class ${targetClass}.`);
-    return javaFileData;
-  }
-
-  const openingBraceIndex = javaFileData.indexOf('{', methodMatch.index);
-  let closingBraceIndex = -1;
-  let braceCount = 1;
-
-  for (let i = openingBraceIndex + 1; i < javaFileData.length; i += 1) {
-    if (javaFileData[i] === '{') {
-      braceCount += 1;
-    } else if (javaFileData[i] === '}') {
-      braceCount -= 1;
-    }
-
-    if (braceCount === 0) {
-      closingBraceIndex = i;
-      break;
-    }
-  }
-
-  if (closingBraceIndex === -1) {
-    console.error(`Could not find closing brace for method ${targetMethod} in class ${targetClass}.`);
-    return javaFileData;
-  }
-
-  const methodBody = javaFileData.slice(openingBraceIndex + 1, closingBraceIndex);
-
-  if (methodBody.includes(codeToInsert.trim())) {
-    return javaFileData;
-  }
-
-  let insertPosition = closingBraceIndex;
-
-  if (lineToInsertAfter) {
-    const lineIndex = methodBody.indexOf(lineToInsertAfter.trim());
-    if (lineIndex !== -1) {
-      insertPosition = openingBraceIndex + 1 + lineIndex + lineToInsertAfter.trim().length;
-    } else {
-      console.error(`Line "${lineToInsertAfter}" not found in method ${targetMethod} of class ${targetClass}.`);
-      return javaFileData;
-    }
-  }
-
-  return `${javaFileData.slice(0, insertPosition)}\n    ${codeToInsert}${javaFileData.slice(insertPosition)}`;
-}
+const {
+  REACT_FONT_MANAGER_JAVA_IMPORT,
+  toArrayBuffer,
+  normalizeString,
+  getProjectFilePath,
+  getFontFamily,
+  getFontFallbackWeight,
+  getFontResFolderPath,
+  getXMLFontId,
+  buildXMLFontObjectEntry,
+  buildXMLFontObject,
+  getAddCustomFontMethodCall,
+  addImportToJavaFile,
+  insertLineInJavaClassMethod,
+  xmlParser,
+  xmlBuilder,
+} = require('../font-assets-helper');
 
 module.exports = function copyAssetsAndroid(assetFiles = [], config, options, fileExt) {
   // If the assets are not fonts, just copy them.
@@ -232,6 +84,9 @@ module.exports = function copyAssetsAndroid(assetFiles = [], config, options, fi
     const isItalic = Boolean((fsSelection & fsSelectionItalicMask) &&
       (macStyle & macStyleItalicMask));
 
+    // Build the font family's map, where each key is the font family name,
+    // and each value is a object containing all the font files related to that
+    // font family.
     const fontFamily = getFontFamily(font.names.fontFamily, font.names.preferredFamily);
     if (!fontFamilyMap[fontFamily]) {
       fontFamilyMap[fontFamily] = {
@@ -246,17 +101,6 @@ module.exports = function copyAssetsAndroid(assetFiles = [], config, options, fi
       weight,
       isItalic,
     });
-  });
-
-  const xmlParser = new XMLParser({
-    ignoreAttributes: false,
-    isArray: tagName => tagName === 'font',
-  });
-
-  const xmlBuilder = new XMLBuilder({
-    format: true,
-    ignoreAttributes: false,
-    suppressEmptyNode: true,
   });
 
   Object.entries(fontFamilyMap).forEach(([fontFamilyName, fontFamilyData]) => {
